Finished API. Need to write unit tests

+++ Backend/ContactsAPI/Dto/ContactRequestDto.cs
@@ -9,9 +9,9 @@ public class ContactRequestDto
     public string LastName { get; set; }
 
     [Required]
-    public int MobilePhoneNumber { get; set; }
+    public double MobilePhoneNumber { get; set; }
 
-    public int HomePhoneNumber { get; set; }
+    public double HomePhoneNumber { get; set; }
     public string Email { get; set; }
 }
