+++ Backend/ContactsAPI/Models/Contacts.cs
@@ -9,10 +9,10 @@ public class Contacts
 	public string LastName { get; set; }
 
 	[Required]
-	public int MobilePhoneNumber { get; set; }
+	public double MobilePhoneNumber { get; set; }
 
 
-    public int HomePhoneNumber { get; set; }
+    public double HomePhoneNumber { get; set; }
     public string Email { get; set; }
 }
